requirements.txt
alembic
asyncpg
aiosqlite
databases
email-validator
fastapi[standard]
fastapi-jwt-auth
httpx
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
pyjwt
pytest
uvicorn
sqlalchemy[asyncio]
requests